#S06P22A306-151 Refactor: Game Api로 받아오는 Data Type 변경

See merge request s06-bigdata-rec-sub2/S06P22A306!23

+++ backend/src/main/java/com/sixteam/ssgame/api/game/entity/MemberGameList.java
@@ -28,12 +28,12 @@ public class MemberGameList {
     private GameInfo gameInfo;
 
     @Column(nullable = false)
-    private Integer memberPlayTime;
+    private Long memberPlayTime;
 
     private Integer memberGameRating;
 
     @Builder
-    public MemberGameList(Member member, GameInfo gameInfo, Integer memberPlayTime, Integer memberGameRating) {
+    public MemberGameList(Member member, GameInfo gameInfo, Long memberPlayTime, Integer memberGameRating) {
         this.member = member;
         this.gameInfo = gameInfo;
         this.memberPlayTime = memberPlayTime;
